requirements/base.txt
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
black==24.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.7
django-filter==24.1
django-phonenumber-field==7.3.0
django-taggit==5.0.1
django-templated-mail==1.1.1
djangorestframework==3.15.0
djangorestframework_simplejwt==5.5.0
djoser==2.2.2
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.7
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
watchfiles==0.21.0
django-celery-beat
django-celery-email
phonenumbers